requirements.txt
fastapi
pydantic
marvin
jinjax
python-dotenv
authlib
pygal
itsdangerous
sqlmodel
psycopg2-binary